refactor: почистила комментарии перед пр
