+++ src/app/globals.scss
@@ -115,5 +115,4 @@
   padding: 0;
   margin: 0;
   list-style: none;
-  // outline: 1px solid violet;
 }
