--- src/components/Contacts/ContactsForm/Input/ClearInputButton/styles.module.scss
@@ -1,42 +1,0 @@
-.input-clear-btn,
-.textarea-clear-btn,
-.promo-input-clear-btn {
-	all: initial;
-	cursor: pointer;
-	position: absolute;
-}
-
-.textarea-clear-btn {
-	position: absolute;
-	top: -6px;
-	right: -6px;
-	background-color: var(--color-bg-white);
-	width: 21px;
-	height: 21px;
-	padding: 1px 0 0 1px;
-	border-radius: 100px;
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	box-shadow: 0px 2px 6.5px 0px rgba(13, 9, 9, 0.15);
-}
-
-.input-clear-btn {
-	position: absolute;
-	top: -6px;
-	right: -6px;
-	background-color: var(--color-bg-white);
-	width: 21px;
-	height: 21px;
-	padding: 1px 0 0 1px;
-	border-radius: 100px;
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	box-shadow: 0px 2px 6.5px 0px rgba(13, 9, 9, 0.15);
-}
-
-.promo-input-clear-btn {
-	top: 42px;
-	right: 15px;
-}
