--- src/components/Contacts/ContactsForm/interface.ts
@@ -1,5 +1,0 @@
-import { FormEvent } from 'react';
-
-export default interface IContactsFormProps {
-	onSubmit: (event: FormEvent<HTMLFormElement>) => void;
-}
