+++ src/components/Input/ClearInputButton/index.tsx
@@ -1,4 +1,4 @@
-import ClearInputIcon from '../../../../../../public/images/icons/ClearInputIcon';
+import ClearInputIcon from '../../../../public/images/icons/ClearInputIcon';
 import styles from './styles.module.scss';
 
 export default function ClearInputButton({
