+++ src/components/Input/ClearInputButton/styles.module.scss
@@ -1,0 +1,23 @@
+.input-clear-btn,
+.textarea-clear-btn,
+.promo-input-clear-btn {
+	all: initial;
+	cursor: pointer;
+	color: var(--color-text-input-active);
+	position: absolute;
+}
+
+.textarea-clear-btn {
+	top: 20px;
+	right: 20px;
+}
+
+.input-clear-btn {
+	top: 18px;
+	right: 15px;
+}
+
+.promo-input-clear-btn {
+	top: 42px;
+	right: 15px;
+}
