+++ src/components/Input/styles.module.scss
@@ -1,0 +1,38 @@
+.promo-form-input {
+	width: 100%;
+	padding: 18px 10px;
+	background-color: var(--color-gray-bg);
+	font-family: var(--involve-font);
+	border: none;
+	border-radius: 10px;
+	transition: all 0.3s ease-out;
+
+	&::placeholder {
+		font-size: var(--text-sm);
+		font-family: var(--involve-font);
+		color: var(--color-input);
+	}
+
+	&:focus {
+		outline: 0;
+		background-color: var(--color-input-active);
+		color: var(--color-text-input-active);
+		transition: all 0.3s ease-out;
+
+		&::placeholder {
+			color: transparent;
+		}
+	}
+}
+
+.promo-form-input-label {
+	margin-right: auto;
+	line-height: 80%;
+}
+
+.promo-form-input-box {
+	position: relative;
+	display: flex;
+	flex-direction: column;
+	row-gap: 10px;
+}
